more crypto testing

+++ util/strings.go
@@ -13,43 +13,43 @@ func SplitN(s, sep string, n int) []string {
 	return strings.SplitN(s, sep, n)
 }
 
-func ParseUint16(s string) (uint16, error) {
+func ParseUint16(s string) (int, error) {
 	x, err := strconv.ParseUint(s, 10, 16)
 	if err != nil {
 		return 0, err
 	}
-	return uint16(x), nil
+	return int(x), nil
 }
 
-func ParseUint32(s string) (uint32, error) {
+func ParseUint32(s string) (int, error) {
 	x, err := strconv.ParseUint(s, 10, 32)
 	if err != nil {
 		return 0, err
 	}
-	return uint32(x), nil
+	return int(x), nil
 }
 
-func ParseUint64(s string) (uint64, error) {
+func ParseUint64(s string) (int, error) {
 	x, err := strconv.ParseUint(s, 10, 64)
 	if err != nil {
 		return 0, err
 	}
-	return x, nil
+	return int(x), nil
 }
 
-func MustParseUint16(s string) uint16 {
+func MustParseUint16(s string) int {
 	x, err := ParseUint16(s)
 	Check(err)
 	return x
 }
 
-func MustParseUint32(s string) uint32 {
+func MustParseUint32(s string) int {
 	x, err := ParseUint32(s)
 	Check(err)
 	return x
 }
 
-func MustParseUint64(s string) uint64 {
+func MustParseUint64(s string) int {
 	x, err := ParseUint64(s)
 	Check(err)
 	return x
